Replace ParameterizedLoaderIT - Parameters cannot be used on classes with JUnit 5
